+++ examples/basic-express-example/README.md
@@ -1,0 +1,20 @@
+## Sync sample application example
+
+## Requirements
+
+- mongodb 
+
+MONGO_CONNECTION_URL environment variable needs to point to mongodb instance
+
+- redis
+
+REDIS_CONNECTION_URL  environment variable needs to point to running redis instance
+
+## Running 
+
+    npm install 
+    npm run start
+
+## Testing
+
+Please refer to documentation for information how to setup sync client.
